// frontend/src/domain/invoice/invoiceModals.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../api/graphql";

export type ServiceItemsVariables = {
  storeId: string;
};

export type ServiceItemsResult = {
  items: ({
  __typename: "ItemConnector";
} & {
  nodes: Array<{
  id: string;
  code: string;
  name: string;
}>;
});
};

export const ServiceItems = {
  query: "query serviceItems($storeId: String!) {\n  items(\n    storeId: $storeId\n    filter: {type: {equalTo: SERVICE}, isActive: true, isVisible: true}\n    page: {first: 100}\n  ) {\n    ... on ItemConnector {\n      __typename\n      nodes {\n        id\n        code\n        name\n      }\n    }\n  }\n}",
} as TypedDocument<ServiceItemsResult, ServiceItemsVariables>;